add; new errors package
